Send all notifications to targetList when gateway
When running as a gateway, send all bucket notifications to every
configured notification target. This enables administrative monitoring
of use of the gateway.

Control for whether notifications are sent or not is based on the
setting of the MINIO_GATEWAY_NOTIFY environment variable. This setting
is a mirror of the notify section of the old config.json file that is no
longer used when in gateway mode. All notification targets configured
will receive all object notifications. There is currently no event filter
mechanism for the gateway generated notifications.